Update product checkout flow
- Add cart add button that lets user add or multiple items at a time
- Remove individual cart add/remove buttons
- Remove column that let users select multiple items

Previously the cart flow would let users either add/remove items
individually as well as having the option to select a checkbox next to
each item, and then add all the checked items. This created two ways to
add items to the checkout and it wasn't clear to the user which way
was correct

The new flow allows users to add new products to the cart in one click
with varying quantities. Once items are in the cart, the user can adjust
the quanitites and the cart will reflect that.